requirements.txt
torch==1.6.0
transformers==3.3.1
sentence-transformers==0.3.8
s3fs==0.5.1
pandas==1.1.2
faiss-cpu==1.6.1
numpy==1.18.5
streamlit==0.62.0
-e .